+++ apps/backoffice-v2/src/pages/MerchantMonitoring/components/MerchantMonitoringTable/MerchantMonitoringTable.tsx
@@ -1,13 +1,16 @@
-import React, { FunctionComponent } from 'react';
-import { TBusinessReports } from '@/domains/business-reports/fetchers';
-import { columns } from '@/pages/MerchantMonitoring/components/MerchantMonitoringTable/columns';
-import { useMerchantMonitoringTableLogic } from '@/pages/MerchantMonitoring/components/MerchantMonitoringTable/hooks/useMerchantMonitoringTableLogic/useMerchantMonitoringTableLogic';
+import type { FunctionComponent } from 'react';
+
 import { UrlDataTable } from '@/common/components/organisms/UrlDataTable/UrlDataTable';
+import { TBusinessReports } from '@/domains/business-reports/fetchers';
+import { useColumns } from '@/pages/MerchantMonitoring/components/MerchantMonitoringTable/columns';
+import { useMerchantMonitoringTableLogic } from '@/pages/MerchantMonitoring/components/MerchantMonitoringTable/hooks/useMerchantMonitoringTableLogic/useMerchantMonitoringTableLogic';
 
 export const MerchantMonitoringTable: FunctionComponent<{
   data: TBusinessReports['data'];
-}> = ({ data }) => {
+  isDemoAccount: boolean;
+}> = ({ data, isDemoAccount }) => {
   const { Cell } = useMerchantMonitoringTableLogic();
+  const columns = useColumns({ isDemoAccount });
 
   return (
     <UrlDataTable