+++ services/workflows-service/src/app.module.ts
@@ -49,6 +49,7 @@ import { RuleEngineModule } from './rule-engine/rule-engine.module';
 import { NotionModule } from '@/notion/notion.module';
 import { SecretsManagerModule } from '@/secrets-manager/secrets-manager.module';
 import { NoteModule } from '@/note/note.module';
+import { MerchantMonitoringModule } from './merchant-monitoring/merchant-monitoring.module';
 
 export const validate = async (config: Record<string, unknown>) => {
   const zodEnvSchema = z
@@ -90,6 +91,7 @@ export const validate = async (config: Record<string, unknown>) => {
     }),
     EventEmitterModule.forRoot(),
     UserModule,
+    MerchantMonitoringModule,
     WorkflowModule,
     WebhooksModule,
     NoteModule,