+++ services/workflows-service/src/workflow/workflow.controller.external.intg.test.ts
@@ -39,7 +39,7 @@ import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.servi
 import { WorkflowControllerExternal } from '@/workflow/workflow.controller.external';
 import { HookCallbackHandlerService } from '@/workflow/hook-callback-handler.service';
 import { DataInvestigationService } from '@/data-analytics/data-investigation.service';
-import { MerchantMonitoringClient } from '@/business-report/merchant-monitoring-client';
+import { MerchantMonitoringClient } from '@/merchant-monitoring/merchant-monitoring.client';
 import { WorkflowEventEmitterService } from '@/workflow/workflow-event-emitter.service';
 import { fetchServiceFromModule, initiateNestApp } from '@/test/helpers/nest-app-helper';
 import { WorkflowTokenRepository } from '@/auth/workflow-token/workflow-token.repository';