+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ballerine/headless-example
 
+## 0.3.98
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.78
+  - @ballerine/workflow-browser-sdk@0.6.99
+
 ## 0.3.97
 
 ### Patch Changes