+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-core
 
+## 0.6.99
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.78
+
 ## 0.6.98
 
 ### Patch Changes